clean up api/data tests
